requirements.txt
flask
sentence-transformers
mysql-connector-python
numpy
faiss-cpu
torch
nltk
python-dotenv
gunicorn
waitress
gevent